Catalog

+++ source_node_final_ecommerce/BE/src/app/controllers/BrandController.js
@@ -1,7 +1,7 @@
-const {
-  CATEGORY_STATUS,
-} = require('../../../../fe/src/constants/productConstant');
+
+const { CATEGORY_STATUSES } = require('../../constants/dbEnum');
 const BrandModel = require('../models/BrandModel');
+const ProductModel = require('../models/ProductModel');
 
 class BrandController {
   // [GET] | api/brands
@@ -145,9 +145,9 @@ class BrandController {
       });
     }
 
-    const image = req.file ? `/uploads/products/${req.file.filename}` : '';
+    const imageUrl = req.file ? req.file.path : '';
     try {
-      const existingBrand = await BrandModel.findOne({ name });
+      const existingBrand = await BrandModel.findOne({ name: { $regex: name, $options: 'i' } });
       if (existingBrand) {
         return res.status(400).json({
           success: false,
@@ -158,7 +158,7 @@ class BrandController {
       const newBrand = new BrandModel({
         name,
         description,
-        image,
+        image_url: imageUrl,
         category_id,
         status,
       });
@@ -181,6 +181,7 @@ class BrandController {
   // [PUT] api/brands/:id/edit
   async update(req, res, next) {
     let { id } = req.params;
+    id = parseInt(id);
     const { name, description, category_id, status } = req.body;
     if (!id) {
       return res.status(400).json({
@@ -196,10 +197,10 @@ class BrandController {
         message: 'Dữ liệu cập nhật không được để trống',
         data: null,
       });
-    }
+    } 
 
     if (req.file) {
-      req.body.image = `/uploads/products/${req.file.filename}`;
+      req.body.image_url = req.file.path;
     }
 
     try {
@@ -217,8 +218,8 @@ class BrandController {
       if (description) brand.description = description;
       if (category_id) brand.category_id = category_id;
       if (status) brand.status = status;
-      if (req.body.image) brand.image = req.body.image;
-      brand.deleted = status === CATEGORY_STATUS.INACTIVE;
+      if (req.body.image_url) brand.image_url = req.body.image_url;
+      brand.deleted = status === 'INACTIVE' ? true : false;
       console.log('Updated brand data:', brand);
       await brand.save();
 
@@ -228,6 +229,7 @@ class BrandController {
         data: brand,
       });
     } catch (error) {
+      console.error('Error updating brand:', error);
       return res.status(500).json({
         success: false,
         message: 'Lỗi server',
@@ -240,7 +242,28 @@ class BrandController {
   // [DELETE] api/brands/:id
   async softDelete(req, res, next) {
     try {
-      const { id } = req.params;
+      let { id } = req.params;
+      
+      if (!id) {
+        return res.status(400).json({
+          success: false,
+          message: 'id không được để trống',
+          data: null,
+        });
+      }
+
+      id = parseInt(id);
+
+      // Kiểm tra xem có sản phẩm nào thuộc thương hiệu này không
+      const hasProducts = await ProductModel.findOne({ brand_id: id }); // + index {brandId:1}
+      if (hasProducts) {
+        return res.status(409).json({
+          success: false,
+          message: 'Không thể xoá: thương hiệu đang có sản phẩm.',
+        });
+      }
+
+
       const c = await BrandModel.delete({ _id: id });
       if (c.deletedCount === 0) {
         return res.status(404).json({
@@ -248,10 +271,10 @@ class BrandController {
           message: 'Không tìm thấy thương hiệu để xóa',
         });
       }
-      await BrandModel.updateOne(
-        { _id: id },
-        { status: CATEGORY_STATUSES.INACTIVE }
-      );
+      // await BrandModel.updateOne(
+      //   { _id: id },
+      //   { status: CATEGORY_STATUSES.INACTIVE }
+      // );
       return res.status(200).json({
         success: true,
         data: id,
@@ -259,7 +282,13 @@ class BrandController {
       });
     } catch (error) {
       console.error(error);
-      next(error);
+      // next(error);
+      return res.status(500).json({
+        success: false,
+        message: 'Lỗi server',
+        data: null,
+        error,
+      });
     }
   }
 }
